requirements.txt
python-telegram-bot==13.7
requests==2.25.1
beautifulsoup4==4.9.3
